backend/pyproject.toml
[project]
name = "trading-system"
version = "0.1.0"
description = "Auto Forex Trader Backend"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2,<5.3",
    "djangorestframework>=3.15.0",
    "drf-spectacular>=0.29.0",
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.0",
    "v20>=3.0.25",
    "boto3>=1.35.0",
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
    "redis>=5.0.0",
    "django-redis>=5.4.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "cryptography>=43.0.0",
    "psutil>=6.1.0",
    "requests>=2.32.0",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.8.0",
    "hypothesis>=6.100.0",
    "hypothesis-django>=0.1.0",
    "factory-boy>=3.3.0",
    "faker>=30.0.0",
    "freezegun>=1.5.0",
    "responses>=0.25.0",
    "fakeredis>=2.26.0",
    "ruff>=0.9.0",
    "ty>=0.0.1a7",
    "pre-commit>=4.0.0",
    "bandit>=1.7.0",
    "interrogate>=1.7.0",
    "types-PyYAML>=6.0.0",
    "types-redis>=4.6.0",
    "types-pyjwt>=1.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--reuse-db",
    "--strict-markers",
    "--strict-config",
    "--import-mode=importlib",
    "-v",
]
testpaths = ["tests"]
markers = [
    "unit: marks tests as unit tests (deselect with '-m \"not unit\"')",
    "integration: Integration tests",
    "slow: Slow running tests",
]
# pytest-xdist configuration for parallel execution
# Use -n auto to automatically detect number of CPUs
# Use --dist loadscope to distribute tests by module for better isolation

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    "build",
    "dist",
    "migrations",
    "*.egg-info",
]

[tool.ruff.lint]
ignore = [
    "E203", # Whitespace before ':'
    "E501", # Line too long (handled by formatter)
]
[tool.ruff.lint.per-file-ignores]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "DJ",  # flake8-django
]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["config", "apps"]
known-third-party = ["django", "rest_framework", "celery", "channels"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "*/asgi.py",
    "*/wsgi.py",
]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.bandit]
exclude_dirs = [".venv", "tests", "migrations"]
skips = ["B101", "B601"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 0
exclude = ["setup.py", "docs", "build", "migrations", "tests"]
verbose = 0
quiet = false
whitelist-regex = []
color = true

[dependency-groups]
dev = [
    "django-stubs>=5.2.8",
    "django-stubs-ext>=5.2.8",
    "pytest-xdist>=3.8.0",
    "types-pyjwt>=1.7.1",
]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
reportMissingImports = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportGeneralTypeIssues = false
reportOptionalMemberAccess = false
reportOptionalSubscript = false
reportPrivateImportUsage = false
exclude = [
    "**/migrations",
    "**/__pycache__",
    "**/node_modules",
    ".venv",
    "tests",
]
